2.3 | Added checks of dependencies before launching programm
